backend/requirements.txt
fastapi==0.103.1
uvicorn==0.23.2
pandas==2.1.0
numpy==1.25.2
scikit-learn==1.3.0
python-multipart==0.0.6
joblib==1.3.2
pyngrok==6.0.0
tensorflow==2.18.0
mne==1.9.0
opencv-python==4.11.0.86
Pillow==11.1.0
